requirements/base.txt
Django
djangorestframework
drf-spectacular
ruff
django-cors-headers